requirements.txt
tensorflow
# tensorflow-addons
tensorflow-probability
numpy
pandas
scikit-learn
opencv-python
Pillow
tqdm
pyyaml
matplotlib
tensorboard
seaborn